Move `TaskItem` and `CustomListRenderer` into their own files
