+++ src/main/java/com/github/tyuioayu/Main.java
@@ -125,61 +125,3 @@ public class Main {
         SwingUtilities.invokeLater(Main::new);
     }
 }
-
-// Task class
-class TaskItem {
-    private String task;
-    private String dueDate;
-    private boolean isCompleted;
-
-    public TaskItem(String task, String dueDate) {
-        this.task = task;
-        this.dueDate = dueDate;
-        this.isCompleted = false;
-    }
-
-    public void toggleCompleted() {
-        isCompleted = !isCompleted;
-    }
-
-    public boolean isCompleted() {
-        return isCompleted;
-    }
-
-    public String getTask() {
-        return task;
-    }
-
-    public void setTask(String task) {
-        this.task = task;
-    }
-
-    public String getDueDate() {
-        return dueDate;
-    }
-
-    public void setDueDate(String dueDate) {
-        this.dueDate = dueDate;
-    }
-
-    @Override
-    public String toString() {
-        return (isCompleted ? "✓ " : "✗ ") + task + " (" + dueDate + ")";
-    }
-}
-
-// Custom list renderer for better UI
-class CustomListRenderer extends DefaultListCellRenderer {
-    private ImageIcon checkIcon = new ImageIcon("check.png");
-    private ImageIcon uncheckIcon = new ImageIcon("uncheck.png");
-
-    @Override
-    public Component getListCellRendererComponent(JList<?> list, Object value, int index, boolean isSelected, boolean cellHasFocus) {
-        JLabel label = (JLabel) super.getListCellRendererComponent(list, value, index, isSelected, cellHasFocus);
-        if (value instanceof TaskItem) {
-            TaskItem task = (TaskItem) value;
-            label.setIcon(task.isCompleted() ? checkIcon : uncheckIcon);
-        }
-        return label;
-    }
-}
